Disable the system.d Emacs service from timing out on startup
Sometimes Emacs takes quite a while to startup on my desktop computer,
when I have a lot of saved buffers for desktop-mode to
restore. Especially when LSP is involved and the corresponding
projects are quite large. Not a big deal for me, but when 90
seconds (the default DefaultTimeoutStartSec value) is exceeded,
system.d was killing the service and starting again - which then
happened recursively. Just disable startup timeouts for the Emacs
service for now to avoid that.